requirements.txt
# 基础依赖
torch>=1.9.0
torchvision>=0.10.0
numpy>=1.19.2
Pillow>=8.3.1
opencv-python>=4.5.3
tqdm>=4.62.3
matplotlib>=3.4.3

# FastAPI依赖
fastapi>=0.68.0
uvicorn>=0.15.0
python-multipart>=0.0.5
pydantic>=1.8.0

# 模型可视化依赖
torchsummary>=1.5.1

# 开发和测试依赖
pytest>=6.2.4
pytest-cov>=2.12.1
black>=21.9b0
flake8>=3.9.2
mypy>=0.910